fix:split a file into component
